+++ spec/rley/parser/ambiguous_parse_spec.rb
@@ -48,7 +48,7 @@ module Rley # Open this namespace to avoid module qualifier prefixes
         factory = ParseWalkerFactory.new
         accept_entry = sentence_result.accepting_entry
         accept_index = sentence_result.chart.last_index
-        @walker = factory.build_walker(accept_entry, accept_index)
+        @walker = factory.build_walker(accept_entry, accept_index, true)
       end
 
       context 'Ambiguous expression' do